Fix: lttng-syscalls.c: `sc_filter` allocation fails
Background
==========
In syscall tracing, all events go througth the same concrete LTTng probe
and are then handled according to the id of the syscall in question (see
`syscall_entry_probe()`).

This means that when enabling the tracing of only one syscall, we need
to filter out all other syscalls to prevent them from ending up in the
trace. This is done using the `sc_filter` bitmap that holds one bit per
syscall.

When a syscall is enabled, we set its corresponding bit in the
`sc_filter` bitmap and bit is tested the next time we want to trace a
syscall.

Note that this filter is only used if when not all syscalls are traced
so it's created on-demand in the `lttng_syscall_filter_enable()`. When
already allocated, it's freed if the user changes his mind and decides
to trace all syscalls.

Issue
=====
In reality, the `sc_filter` is, in fact, never allocated. In the
`lttng_syscall_filter_enable()` function, the `sc_filter` bitmap is
freed if we realize that we fail to find the syscall name in the
`sc_table` syscall table build statically.

During my tests, I witnessed that this failed all the time because of an
issue with the comparison function `get_syscall_nr()`. This function
erroneously trims the name of the syscall to remove the common prefix
"syscall_entry_" before passing it to `strcmp()`.

So we end up passing the following to `strcmp()` for the `openat`
syscall:
  strcmp (cs=0xffffffff821ad7b6 "syscall_entry_openat",
          ct=0xffffffff821ad7c4 "openat")

The name that we deal with at this point of the processing is the full
name that includes the prefix as we received it from the
`lttng_event_desc` structure. The work of ignoring the prefix is done in
`lttng_desc_match_enabler()`.

This problem with the comparison makes it so that this function always
returns that syscalls are never found in the table.

So, we end up in a situation as if all syscalls are enabled. So syscall
probes are called all the time even if that syscall was not enabled by
the user.

Enabling specific syscall still works because the `lttng_event::enabled`
field is checked in the probe of each syscall so we exit early in the
probe when disabled.

The issue is that we are doing work that will be discarded in the fast
path.

Solution
========
Get rid of the prefix trimming by comparing the both strings directly.

Now that we are successfully comparing the name we end up creating and
using the `sc_filter` bitmap and thus avoiding to call the event probe
of disabled syscalls.

This avoids a whole bunch of work in the tracing fast path.

Note
====
The issue with `get_syscall_nr()` is also present in
`get_compat_syscall_nr()`, so the fix is the same.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Change-Id: I9c2b9c91ee9c668784533bfb97faa9b4909913b6